Create code.ino
This project consists of an arduino uno, a RFID reader (RC522) and two RFID tags. The purpose of this project is to built a system that will give permission for an operation if the wright tag passes the RFID reader. Analytical documentation presented in greek is uploaded in https://drive.google.com/file/d/1XjBXu3V32Qq0wZYJunxhfVCewNNYj_19/view?usp=sharing.